Added ability to upload to blob from save func
From save_Read_result function, adding a new argument that allows the
user to save directly to a Azure blob container. This saves having to
write to a local file and re-read to upload to azure.

Also fixed a bug where only the last page of the pdf was being extracted
and saved.